Move webaudio files to their own folder Signed-off-by: Ashwin Naren <arihant2math@gmail.com>
