+++ components/script/dom/audio/baseaudiocontext.rs
@@ -22,12 +22,21 @@ use servo_media::{ClientContextId, ServoMedia};
 use uuid::Uuid;
 
 use crate::conversions::Convert;
-use crate::dom::analysernode::AnalyserNode;
-use crate::dom::audiobuffer::AudioBuffer;
-use crate::dom::audiobuffersourcenode::AudioBufferSourceNode;
-use crate::dom::audiodestinationnode::AudioDestinationNode;
-use crate::dom::audiolistener::AudioListener;
-use crate::dom::audionode::MAX_CHANNEL_COUNT;
+use crate::dom::audio::analysernode::AnalyserNode;
+use crate::dom::audio::audiobuffer::AudioBuffer;
+use crate::dom::audio::audiobuffersourcenode::AudioBufferSourceNode;
+use crate::dom::audio::audiodestinationnode::AudioDestinationNode;
+use crate::dom::audio::audiolistener::AudioListener;
+use crate::dom::audio::audionode::MAX_CHANNEL_COUNT;
+use crate::dom::audio::biquadfilternode::BiquadFilterNode;
+use crate::dom::audio::channelmergernode::ChannelMergerNode;
+use crate::dom::audio::channelsplitternode::ChannelSplitterNode;
+use crate::dom::audio::constantsourcenode::ConstantSourceNode;
+use crate::dom::audio::gainnode::GainNode;
+use crate::dom::audio::iirfilternode::IIRFilterNode;
+use crate::dom::audio::oscillatornode::OscillatorNode;
+use crate::dom::audio::pannernode::PannerNode;
+use crate::dom::audio::stereopannernode::StereoPannerNode;
 use crate::dom::bindings::callback::ExceptionHandling;
 use crate::dom::bindings::cell::DomRefCell;
 use crate::dom::bindings::codegen::Bindings::AnalyserNodeBinding::AnalyserOptions;
@@ -53,18 +62,9 @@ use crate::dom::bindings::num::Finite;
 use crate::dom::bindings::refcounted::Trusted;
 use crate::dom::bindings::reflector::DomGlobal;
 use crate::dom::bindings::root::{DomRoot, MutNullableDom};
-use crate::dom::biquadfilternode::BiquadFilterNode;
-use crate::dom::channelmergernode::ChannelMergerNode;
-use crate::dom::channelsplitternode::ChannelSplitterNode;
-use crate::dom::constantsourcenode::ConstantSourceNode;
 use crate::dom::domexception::{DOMErrorName, DOMException};
 use crate::dom::eventtarget::EventTarget;
-use crate::dom::gainnode::GainNode;
-use crate::dom::iirfilternode::IIRFilterNode;
-use crate::dom::oscillatornode::OscillatorNode;
-use crate::dom::pannernode::PannerNode;
 use crate::dom::promise::Promise;
-use crate::dom::stereopannernode::StereoPannerNode;
 use crate::realms::InRealm;
 use crate::script_runtime::CanGc;
