+++ components/script/dom/audio/mediastreamaudiodestinationnode.rs
@@ -8,8 +8,8 @@ use servo_media::ServoMedia;
 use servo_media::audio::node::AudioNodeInit;
 use servo_media::streams::MediaStreamType;
 
-use crate::dom::audiocontext::AudioContext;
-use crate::dom::audionode::{AudioNode, AudioNodeOptionsHelper};
+use crate::dom::audio::audiocontext::AudioContext;
+use crate::dom::audio::audionode::{AudioNode, AudioNodeOptionsHelper};
 use crate::dom::bindings::codegen::Bindings::AudioNodeBinding::{
     AudioNodeOptions, ChannelCountMode, ChannelInterpretation,
 };
