+++ components/script/dom/audio/mediastreamtrackaudiosourcenode.rs
@@ -6,8 +6,8 @@ use dom_struct::dom_struct;
 use js::rust::HandleObject;
 use servo_media::audio::node::AudioNodeInit;
 
-use crate::dom::audiocontext::AudioContext;
-use crate::dom::audionode::AudioNode;
+use crate::dom::audio::audiocontext::AudioContext;
+use crate::dom::audio::audionode::AudioNode;
 use crate::dom::bindings::codegen::Bindings::MediaStreamTrackAudioSourceNodeBinding::{
     MediaStreamTrackAudioSourceNodeMethods, MediaStreamTrackAudioSourceOptions,
 };
