+++ components/script/dom/audio/mod.rs
@@ -1,0 +1,31 @@
+/* This Source Code Form is subject to the terms of the Mozilla Public
+ * License, v. 2.0. If a copy of the MPL was not distributed with this
+ * file, You can obtain one at https://mozilla.org/MPL/2.0/. */
+
+pub(crate) mod analysernode;
+pub(crate) mod audiobuffer;
+pub(crate) mod audiobuffersourcenode;
+pub(crate) mod audiocontext;
+pub(crate) mod audiodestinationnode;
+pub(crate) mod audiolistener;
+pub(crate) mod audionode;
+pub(crate) mod audioparam;
+pub(crate) mod audioscheduledsourcenode;
+pub(crate) mod audiotrack;
+pub(crate) mod audiotracklist;
+pub(crate) mod baseaudiocontext;
+pub(crate) mod biquadfilternode;
+pub(crate) mod channelmergernode;
+pub(crate) mod channelsplitternode;
+pub(crate) mod constantsourcenode;
+pub(crate) mod gainnode;
+pub(crate) mod iirfilternode;
+pub(crate) mod mediaelementaudiosourcenode;
+pub(crate) mod mediastreamaudiodestinationnode;
+pub(crate) mod mediastreamaudiosourcenode;
+pub(crate) mod mediastreamtrackaudiosourcenode;
+pub(crate) mod offlineaudiocompletionevent;
+pub(crate) mod offlineaudiocontext;
+pub(crate) mod oscillatornode;
+pub(crate) mod pannernode;
+pub(crate) mod stereopannernode;
