+++ components/script/dom/audio/oscillatornode.rs
@@ -15,10 +15,10 @@ use servo_media::audio::oscillator_node::{
 use servo_media::audio::param::ParamType;
 
 use crate::conversions::Convert;
-use crate::dom::audionode::AudioNodeOptionsHelper;
-use crate::dom::audioparam::AudioParam;
-use crate::dom::audioscheduledsourcenode::AudioScheduledSourceNode;
-use crate::dom::baseaudiocontext::BaseAudioContext;
+use crate::dom::audio::audionode::AudioNodeOptionsHelper;
+use crate::dom::audio::audioparam::AudioParam;
+use crate::dom::audio::audioscheduledsourcenode::AudioScheduledSourceNode;
+use crate::dom::audio::baseaudiocontext::BaseAudioContext;
 use crate::dom::bindings::codegen::Bindings::AudioNodeBinding::{
     ChannelCountMode, ChannelInterpretation,
 };
