+++ components/script/dom/mod.rs
@@ -217,23 +217,12 @@ pub(crate) mod abstractrange;
 pub(crate) mod abstractworker;
 pub(crate) mod abstractworkerglobalscope;
 pub(crate) mod activation;
-pub(crate) mod analysernode;
 pub(crate) mod animationevent;
 pub(crate) mod attr;
-pub(crate) mod audiobuffer;
-pub(crate) mod audiobuffersourcenode;
-pub(crate) mod audiocontext;
-pub(crate) mod audiodestinationnode;
-pub(crate) mod audiolistener;
-pub(crate) mod audionode;
-pub(crate) mod audioparam;
-pub(crate) mod audioscheduledsourcenode;
-pub(crate) mod audiotrack;
-pub(crate) mod audiotracklist;
-pub(crate) mod baseaudiocontext;
+pub(crate) mod audio;
+pub(crate) use self::audio::*;
 pub(crate) mod beforeunloadevent;
 pub(crate) mod bindings;
-pub(crate) mod biquadfilternode;
 pub(crate) mod blob;
 #[cfg(feature = "bluetooth")]
 #[allow(clippy::module_inception)]
@@ -247,8 +236,6 @@ pub(crate) mod canvaspattern;
 #[allow(dead_code)]
 pub(crate) mod canvasrenderingcontext2d;
 pub(crate) mod cdatasection;
-pub(crate) mod channelmergernode;
-pub(crate) mod channelsplitternode;
 pub(crate) mod characterdata;
 pub(crate) mod client;
 pub(crate) mod clipboard;
@@ -258,7 +245,6 @@ pub(crate) mod closeevent;
 pub(crate) mod comment;
 pub(crate) mod compositionevent;
 pub(crate) mod console;
-pub(crate) mod constantsourcenode;
 pub(crate) mod cookiestore;
 pub(crate) mod countqueuingstrategy;
 mod create;
@@ -337,7 +323,6 @@ pub(crate) mod fontface;
 pub(crate) mod fontfaceset;
 pub(crate) mod formdata;
 pub(crate) mod formdataevent;
-pub(crate) mod gainnode;
 pub(crate) mod gamepad;
 pub(crate) mod gamepadbutton;
 pub(crate) mod gamepadbuttonlist;
@@ -435,7 +420,6 @@ pub(crate) mod idbopendbrequest;
 pub(crate) mod idbrequest;
 pub(crate) mod idbtransaction;
 pub(crate) mod idbversionchangeevent;
-pub(crate) mod iirfilternode;
 pub(crate) mod imagebitmap;
 pub(crate) mod imagebitmaprenderingcontext;
 pub(crate) mod imagedata;
@@ -446,7 +430,6 @@ pub(crate) mod keyboardevent;
 pub(crate) mod location;
 pub(crate) mod mediadeviceinfo;
 pub(crate) mod mediadevices;
-pub(crate) mod mediaelementaudiosourcenode;
 pub(crate) mod mediaerror;
 pub(crate) mod mediafragmentparser;
 pub(crate) mod medialist;
@@ -455,10 +438,7 @@ pub(crate) mod mediaquerylist;
 pub(crate) mod mediaquerylistevent;
 pub(crate) mod mediasession;
 pub(crate) mod mediastream;
-pub(crate) mod mediastreamaudiodestinationnode;
-pub(crate) mod mediastreamaudiosourcenode;
 pub(crate) mod mediastreamtrack;
-pub(crate) mod mediastreamtrackaudiosourcenode;
 pub(crate) mod messagechannel;
 pub(crate) mod messageevent;
 #[allow(dead_code)]
@@ -478,16 +458,12 @@ pub(crate) mod nodeiterator;
 #[allow(dead_code)]
 pub(crate) mod nodelist;
 pub(crate) mod notification;
-pub(crate) mod offlineaudiocompletionevent;
-pub(crate) mod offlineaudiocontext;
 pub(crate) mod offscreencanvas;
 pub(crate) mod offscreencanvasrenderingcontext2d;
-pub(crate) mod oscillatornode;
 pub(crate) mod pagetransitionevent;
 pub(crate) mod paintrenderingcontext2d;
 pub(crate) mod paintsize;
 pub(crate) mod paintworkletglobalscope;
-pub(crate) mod pannernode;
 pub(crate) mod path2d;
 pub(crate) mod performance;
 #[allow(dead_code)]
@@ -559,7 +535,6 @@ pub(crate) mod servointernals;
 pub(crate) mod servoparser;
 pub(crate) mod shadowroot;
 pub(crate) mod staticrange;
-pub(crate) mod stereopannernode;
 pub(crate) mod storage;
 pub(crate) mod storageevent;
 pub(crate) mod stylepropertymapreadonly;
